Added 'refersToClassObject' function
In 'A.B' 'A' may refer to class descriptor or class object descriptor
(depends on B, it may be nested class in A; class, property or function in class object).
REFERENCE_TARGET is always class descriptor, but navigation (class or class object) is determined on context ('B').

Fixed 'extract function' test (qualifiedClassObject.kt).