(TODO test diagnostic somehow?) Keep built-in metadata for reflection interfaces
Reflection interfaces (traits in kotlin.reflect.* in core/builtins/) are now
fully considered as built-ins and can be accessed via KotlinBuiltIns. This
increases runtime size by ~20kb, but only because KotlinBuiltIns is static and
is therefore constructed only via resource loading from the compiler classpath
at the moment. As soon as it's possible to inject KotlinBuiltIns to the
particular resolution process, the metadata on JVM will be loaded via standard
annotation mechanism and wasted runtime space will be reclaimed